+++ src/cone/app/browser/layout.py
@@ -86,6 +86,7 @@ class ViewSettingsAction(LinkAction):
 class LogoutAction(LinkAction):
     text = _('logout', default='Logout')
     icon = 'ion-log-out'
+    action = 'logout:NONE:NONE'
     path_action = ''
     path_event = ''
 
@@ -95,7 +96,7 @@ class LogoutAction(LinkAction):
 
     @property
     def path(self):
-        return self.request.application_url
+        return '/'
 
 
 personal_tools = odict()